# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Geoindicadores",
        "description": "Projeto onde armazena indicadores ambientais para análise",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        },
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/variaveis": {
            "get": {
                "tags": [
                    "Variável"
                ],
                "summary": "Traz a lista de Variáveis",
                "description": "Retorna a lista de Variáveis",
                "operationId": "034dfdad1c9e95361d9eb278220df080",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Um exemplo de resultado",
                                        "value": [
                                            {
                                                "id": "integer",
                                                "indicador_id": "integer",
                                                "variavel": {
                                                    "id": "integer",
                                                    "codigo": "string",
                                                    "nome": "string",
                                                    "departamento": {
                                                        "id": "integer",
                                                        "nome": "string",
                                                        "sigla": "string"
                                                    }
                                                },
                                                "tipo_dado": {
                                                    "id": "integer",
                                                    "tipo": "string"
                                                },
                                                "fonte": {
                                                    "id": "integer",
                                                    "sigla": "string",
                                                    "nome": "string"
                                                },
                                                "metadados": {
                                                    "id": "integer",
                                                    "tipo_medida_id": "string",
                                                    "serie_historica_inicio": "string",
                                                    "serie_historica_fim": "string",
                                                    "serie_historica_ativo": "integer",
                                                    "nota_tecnica": "string",
                                                    "organizacao": "string",
                                                    "observação": "string"
                                                }
                                            },
                                            {
                                                "id": "integer",
                                                "indicador_id": "integer",
                                                "variavel": {
                                                    "id": "integer",
                                                    "codigo": "string",
                                                    "nome": "string",
                                                    "departamento": {
                                                        "id": "integer",
                                                        "nome": "string",
                                                        "sigla": "string"
                                                    },
                                                    "tipo_dado": {
                                                        "id": "integer",
                                                        "tipo": "string"
                                                    },
                                                    "fonte": {
                                                        "id": "integer",
                                                        "sigla": "string",
                                                        "nome": "string"
                                                    },
                                                    "metadados": {
                                                        "id": "integer",
                                                        "tipo_medida_id": "string",
                                                        "serie_historica_inicio": "string",
                                                        "serie_historica_fim": "string",
                                                        "serie_historica_ativo": "integer",
                                                        "nota_tecnica": "string",
                                                        "organizacao": "string",
                                                        "observação": "string"
                                                    }
                                                }
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não Encontrado"
                    }
                }
            }
        },
        "/indicador/{indicador_id}/variaveis": {
            "get": {
                "tags": [
                    "Variável"
                ],
                "summary": "Traz a lista de variáveis de um indicador",
                "description": "Retorna a lista de Variáveis de um indicador",
                "operationId": "04118b3ec6863d106a28fe956f161514",
                "parameters": [
                    {
                        "name": "indicador_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Um exemplo de resultado",
                                        "value": [
                                            {
                                                "id": "integer",
                                                "indicador_id": "integer",
                                                "variavel": {
                                                    "id": "integer",
                                                    "codigo": "string",
                                                    "nome": "string",
                                                    "departamento": {
                                                        "id": "integer",
                                                        "nome": "string",
                                                        "sigla": "string"
                                                    },
                                                    "tipo_dado": {
                                                        "id": "integer",
                                                        "tipo": "string"
                                                    },
                                                    "fonte": {
                                                        "id": "integer",
                                                        "sigla": "string",
                                                        "nome": "string"
                                                    },
                                                    "metadados": {
                                                        "id": "integer",
                                                        "tipo_medida_id": "string",
                                                        "serie_historica_inicio": "string",
                                                        "serie_historica_fim": "string",
                                                        "nota_tecnica": "string",
                                                        "organizacao": "string",
                                                        "observação": "string"
                                                    }
                                                }
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não Encontrado"
                    }
                }
            }
        }
    }
}